+++ CHANGE_LOG.md
@@ -9,6 +9,19 @@ implemented
 implemented 
     update hall booking UI
     - kamalesh   
+implemented   
+    added the feature Gmail verification using google-oauth  
+    modified the users table added two new columns "email_otp, email_otp_expires"  
+    - DHARUNKUMAR S    
+implemented  
+    login and register pages  
+    - kabil  
+implemented  
+    updated landing page UI  
+    - loki  
+implemented
+    update donation UI
+    -kamalesh  
 Implemented  
     Update pooja booking UI  
     - Joyal Augustin